feat(collection-tree): migrate some test case to path.test.ts file
